+++ ground/src/main/java/com/google/android/ground/persistence/local/LocalDataStoreModule.kt
@@ -42,9 +42,7 @@ abstract class LocalDataStoreModule {
   abstract fun offlineAreaStore(store: RoomOfflineAreaStore): LocalOfflineAreaStore
   @Binds
   @Singleton
-  abstract fun submissionStore(
-    store: RoomSubmissionMutationStore
-  ): LocalSubmissionMutationStore
+  abstract fun submissionStore(store: RoomSubmissionMutationStore): LocalSubmissionMutationStore
   @Binds @Singleton abstract fun surveyStore(store: RoomSurveyStore): LocalSurveyStore
   @Binds @Singleton abstract fun tileSetStore(store: RoomTileSetStore): LocalTileSetStore
   @Binds @Singleton abstract fun userStore(store: RoomUserStore): LocalUserStore
