+++ ground/src/main/java/com/google/android/ground/persistence/local/room/converter/ConverterExt.kt
@@ -35,7 +35,7 @@ import com.google.android.ground.model.task.Option
 import com.google.android.ground.model.task.Task
 import com.google.android.ground.persistence.local.LocalDataConsistencyException
 import com.google.android.ground.persistence.local.room.entity.*
-import com.google.android.ground.persistence.local.room.models.*
+import com.google.android.ground.persistence.local.room.fields.*
 import com.google.android.ground.persistence.local.room.relations.JobEntityAndRelations
 import com.google.android.ground.persistence.local.room.relations.SurveyEntityAndRelations
 import com.google.android.ground.persistence.local.room.relations.TaskEntityAndRelations
