+++ ground/src/main/java/com/google/android/ground/persistence/local/room/dao/SubmissionDao.kt
@@ -18,7 +18,7 @@ package com.google.android.ground.persistence.local.room.dao
 import androidx.room.Dao
 import androidx.room.Query
 import com.google.android.ground.persistence.local.room.entity.SubmissionEntity
-import com.google.android.ground.persistence.local.room.models.EntityState
+import com.google.android.ground.persistence.local.room.fields.EntityState
 import io.reactivex.Maybe
 import io.reactivex.Single
