+++ ground/src/main/java/com/google/android/ground/persistence/local/room/entity/OfflineAreaEntity.kt
@@ -19,7 +19,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.google.android.ground.model.basemap.OfflineArea
-import com.google.android.ground.persistence.local.room.models.OfflineAreaEntityState
+import com.google.android.ground.persistence.local.room.fields.OfflineAreaEntityState
 
 /** Represents a [OfflineArea] in the local data store. */
 @Entity(tableName = "offline_base_map")
