+++ ground/src/main/java/com/google/android/ground/persistence/local/room/entity/SubmissionEntity.kt
@@ -17,7 +17,7 @@ package com.google.android.ground.persistence.local.room.entity
 
 import androidx.room.*
 import com.google.android.ground.model.submission.Submission
-import com.google.android.ground.persistence.local.room.models.EntityState
+import com.google.android.ground.persistence.local.room.fields.EntityState
 
 /** Representation of a [Submission] in local db. */
 @Entity(
