+++ ground/src/main/java/com/google/android/ground/persistence/local/room/entity/SubmissionMutationEntity.kt
@@ -17,8 +17,8 @@ package com.google.android.ground.persistence.local.room.entity
 
 import androidx.room.*
 import com.google.android.ground.model.mutation.SubmissionMutation
-import com.google.android.ground.persistence.local.room.models.MutationEntitySyncStatus
-import com.google.android.ground.persistence.local.room.models.MutationEntityType
+import com.google.android.ground.persistence.local.room.fields.MutationEntitySyncStatus
+import com.google.android.ground.persistence.local.room.fields.MutationEntityType
 import java.util.*
 
 /** Representation of a [SubmissionMutation] in local data store. */
