+++ ground/src/main/java/com/google/android/ground/persistence/local/room/fields/MutationEntitySyncStatus.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.ground.persistence.local.room.models;
+package com.google.android.ground.persistence.local.room.fields;
 
 import static java8.util.J8Arrays.stream;
