+++ ground/src/main/java/com/google/android/ground/repository/MutationRepository.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Google LLC
+ * Copyright (c) 2023 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.persistence.local.room.models
 
-enum class GeometryType {
-  POLYGON,
-  POINT,
-}
+package com.google.android.ground.repository
+
+class MutationRepository() {}
